+++ src/lib/modules/auth/index.ts
@@ -3,5 +3,8 @@ import type { Module } from '~/types'
 export const AuthModule: Module = {
 	description: 'Auth (User authentication and authorization)',
 	npmDependencies: [{ '@rubriclab/auth': '*' }],
-	moduleDependencies: ['DatabaseModule']
+	moduleDependencies: ['DatabaseModule'],
+	env: {
+		URL: 'string'
+	}
 }