+++ src/lib/modules/auth/src/app/(auth)/signin/magiclink/sent/page.tsx
@@ -1,0 +1,3 @@
+export default function MagicLink() {
+	return <>Check your inbox!</>
+}