+++ src/lib/modules/email/index.ts
@@ -3,5 +3,8 @@ import type { Module } from '~/types'
 export const EmailModule: Module = {
 	description: '',
 	npmDependencies: [{ nodemailer: '^6.7.2' }],
-	npmDevDependencies: [{ '@types/nodemailer': '^6.4.4' }]
+	npmDevDependencies: [{ '@types/nodemailer': '^6.4.4' }],
+	env: {
+		RESEND_API_KEY: 'string'
+	}
 }