+++ src/lib/modules/email/src/lib/email.ts
@@ -1,0 +1,6 @@
+import { Resend } from 'resend'
+import { env } from '~/env'
+
+export const {
+	emails: { send }
+} = new Resend(env.RESEND_API_KEY)